requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.5
django-taggit==6.1.0
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.3
python-dotenv==1.1.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
